requirements.txt
requests==2.31.0
CairoSVG==2.7.1
chess==1.10.0
stockfish==3.28.0
python-telegram-bot[job-queue,webhooks] --pre
zstandard==0.21.0
numpy==1.24.2
pandas==1.5.3
redis==5.0.1
urllib3==2.0.4
Pillow==10.0.0
tqdm==4.66.1
